+++ TombEngine/Game/gui.cpp
@@ -484,6 +484,7 @@ namespace TEN::Gui
 			Windowed,
 			ShadowType,
 			Caustics,
+			Decals,
 			Antialiasing,
 			AmbientOcclusion,
 			HighFramerate,
@@ -534,6 +535,11 @@ namespace TEN::Gui
 				CurrentSettings.Configuration.EnableCaustics = !CurrentSettings.Configuration.EnableCaustics;
 				break;
 
+			case DisplaySettingsOption::Decals:
+				SoundEffect(SFX_TR4_MENU_CHOOSE, nullptr, SoundEnvironment::Always);
+				CurrentSettings.Configuration.EnableDecals = !CurrentSettings.Configuration.EnableDecals;
+				break;
+
 			case DisplaySettingsOption::Antialiasing:
 				SoundEffect(SFX_TR4_MENU_CHOOSE, nullptr, SoundEnvironment::Always);
 
@@ -574,6 +580,7 @@ namespace TEN::Gui
 
 			case DisplaySettingsOption::ShadowType:
 				SoundEffect(SFX_TR4_MENU_CHOOSE, nullptr, SoundEnvironment::Always);
+
 				if (CurrentSettings.Configuration.ShadowType == ShadowMode::All)
 					CurrentSettings.Configuration.ShadowType = ShadowMode::None;
 				else
@@ -584,6 +591,11 @@ namespace TEN::Gui
 			case DisplaySettingsOption::Caustics:
 				SoundEffect(SFX_TR4_MENU_CHOOSE, nullptr, SoundEnvironment::Always);
 				CurrentSettings.Configuration.EnableCaustics = !CurrentSettings.Configuration.EnableCaustics;
+				break;
+
+			case DisplaySettingsOption::Decals:
+				SoundEffect(SFX_TR4_MENU_CHOOSE, nullptr, SoundEnvironment::Always);
+				CurrentSettings.Configuration.EnableDecals = !CurrentSettings.Configuration.EnableDecals;
 				break;
 
 			case DisplaySettingsOption::Antialiasing: